Always keep the display name when deduplicating non-authorized names
The incoming list of agent names is processed to remove duplicates
prior to being saved, so a set like this:

  names: [{primary_name: "hello", authorized: true, is_display_name: true},
          {primary_name: "duplicate", authorized: false, is_display_name: false},
          {primary_name: "duplicate", authorized: false, is_display_name: false}]

Is collapsed to:

  names: [{primary_name: "hello", authorized: true, is_display_name: true},
          {primary_name: "duplicate", authorized: false, is_display_name: false}]

This can cause problems if you see input like this:

  names: [{primary_name: "hello", authorized: true,  is_display_name: false},
          {primary_name: "hello", authorized: false, is_display_name: true}]

The deduplication code just naively takes whichever name it sees
first, but in this case the duplicate was marked as the display name.
Collapsing to:

  names: [{primary_name: "hello", authorized: true,  is_display_name: false}]

leaves us without a display name at all, and then viewing the
resulting agent record throws an error.

This commit refines the behavior to the following:

  * We always keep the authorized name

  * If two names are the same, we keep whichever was marked as the
    display name (if neither was, we can choose either)

  * If a duplicate of the authorized name was marked as the display
    name, the authorized name becomes the display name