Reusable subroutines for handling RecordNotFound (and similar) errors
The same code for setting up and rendering the shared/not_found template is repeated in most of the PUI controllers. This commit moves that to new methods in ApplicationController, cleaning up the code, and allowing plug-ins to override just the not-found response, instead of the entire method for handling core routes.