Change call to embedded search template to use "agent_uris" field
This will make the "Linked Records" section when viewing an agent in the staff interface more accurate, in that it will list only those records (resources, archival objects, etc) which are actually linked to the agent, and not any which are linked to a duplicate agent with the same title. Same chagen as for subjects, except the "agent_uris" Solr field already exists.