ANW-1020. uniq_ao_pos index constraint violation fix
Fix for a possible database constraint violation of the "uniq_ao_pos"
index in the "archival_object" table may occur when editing and saving
and archival object.

In the "backend/app/model/mixins/tree_nodes.rb" file, there is a
distinction between "logical position" and "physical position" of an
archival object. The "physical position" is the actual value in the
"position" field in the "archival_object" table in the database. This
value contains "gaps" to enable insertion of elements without having to
change all the other elements to preserve the order of the object in the
list. The "logical position" is the "gapless" position of the object
in the list, used by the API.

When generating the JSON object from the database in the
"sequel_to_jsonmodel" method, the "physical position" retrieved from the
database is stored in the "json.position" field. This is replaced in the
method by the "logical position", and the logical position is what is
returned by the API. When the object is updated, the "json.position"
field returned from the form still has the logical position. In the
"update_from_json" method in "tree_nodes.rb", an intermediate result
is saved to the database, where the JSON contains the logical position,
not the physical position.

We've seen occur as follows. In our database, the "physical position"
of Archival Object #538301 is "501500", which has a
"logical position" of "500". When attempting to save Archival Object
"500" (the logical positon), which conflicts with
Archival Object #537801, which has a physical position of "500".

This problem only affects specific records because of the gaps in the
physical position. As more records get added to the system, a
logical/physical position conflict becomes more likely.

The implemented fix is to reset the "json.position" field to the
physical position before writing the object to the database in the
"update_from_json" method of "tree_nodes.rb". The "json.position" is
reset back to "logical position" after writing to the database,  as
this appears to be expected by tests such as
"backend/spec/component_transfer_spec.rb" (specifically the "doesn't
break node sequencing" test).

A minor wrinkle is that when the "set_position_in_list" method is run,
the "position" field for the archival object in the database will be
modified, because the "attempt_set_position_in_list" method thinks the
position is already occupied (which it is, but by the object being
updated). The existing code then moves the record into a "gap" in the
physical positions which is not necessary. Adding a check which simply
skips any position change if we are already at the correct logical
position.

https://archivesspace.atlassian.net/browse/ANW-1020