Work around for search issues with stemming & wildcards
Wildcards interact badly with stemming because the wildcard causes
query analysis to be skipped, so stemming isn't applied to the query.

This manifests in real data when you typeahead for "agency*" and get
no matches.  That term is stemmed to "agenc".

Try to minimise the weird effects of this by searching for the
non-wildcard version as well.  The real solution here is to stop using
wildcards and use an ngram field instead.