Move Sequence generation outside of the current transaction
I realise that this class is deprecated, but some plugins do use this
class and the semantics aren't quite right!

The Sequence class is supposed to issue a stream of integers without
repeats.  The transaction retry logic in the original implementation
doesn't really make sense: because MySQL uses repeatable reads, and
because the call to Sequence.get is running in the context of the
request's database transaction, retrying in a loop when a duplicate is
hit is never going to give different results.

To get the semantics we want, the Sequence-getting logic needs to run
in a separate transaction that gets committed even if we rollback the
requesting transaction.  This change achieves this by running the
sequence generator in a background thread.

Ultimately maybe this whole thing gets deleted and plugins have to
pull this in if they need it, but having this example here might at
least be helpful when the time comes.