Browse search prefs final
This is the final branch for the work done by @sem42198

A refactor of how search views in the staff interface which resolves
a range of tickets related to consistency and which columns should
show up in which places. Additionally, allows for more flexibility
for setting which columns should show up for a given record type and
how search results should be sorted by default (either in preferences
or in a config file). Some new options for columns are included also
such as dates and extents.