Updates for pui search
Add fq for types pui to all pui searches, this eliminates duplicates results
with record types using inheritance (i.e. where there are 2 published records
but only 1 with type pui, commonly this applies to archival objects).

Add types pui to tree and scroll docs so the pui can continue to access them.
This is required for the Collection Organization functionality to work.

Update all record types to include a uri. This is for consistency and also
so that the records_for_uris method can always search by uri except when
using a pui id (in that case use the id field as before). This is also
required for Collection Organization as it uses both forms (id and uri).

Cleanup unused cm_uri for Collection Management.