ANW-851: Fix temporary location lost on update
The disabled input means params are not sent to the controller, so
the data is lost on update. Instead, if a temporary location flag
is submitted, check for a value in the params or for a pre-existing
value in the location record. Also ensure that when a temporary
location is selected the flag is checked.