ANW-147 Merging extent controlled values loses related items
Related items relies on search, however the indexer was not being
updated for records "related" to the referenced extents. An active
association mixin has been added to make it convenient to determine
the parent record of the extent and bump the mtime for re-indexing.

The mixin could also be applied to other records types and used for
refactoring in the future where if / case statements are being used
to find an associated record by non-null id.

There is more detail on use cases: ANW-1278
There is also some cleanup for the gnarly touch records mixin spec.