Handle relative paths for `AppConfig[:plugins_directory]`
The default value for the `AppConfig[:plugins_directory]` is:

     # AppConfig[:plugins_directory] = "plugins"

but if you uncomment that line in the standard config.rb, it fails to
find your plugins.  Someone hit that here:

https://groups.google.com/g/archivesspace/c/22bOn1sou0s

The `ASUtils.plugin_base_directory` method was mistakenly interpreting
relative paths against the backend directory, rather than the
ArchivesSpace base directory.