ANW-1347 Fix setting language preferences in staff interface
I18n cache optimization did not include the locale, resulting in
cache hits irrespective of the locale, based on first encountered.

This prefixes the cache entries with the locale to restore the
original functionality. Also supports resetting the cache which
may be useful in the future and updates the locale switching code
to mirror Rails documentation (around_action and with_locale).