Be ready for the hash passed to display_audit_info having different shapes
display_audit_info can be called from a view or from a browse (and who knows
where else in the future). If hash is from a solr result - as is the case
with a call from browse, then hash["ark_name'] is an array with the current
ARK in the first position. If it is from a JSONModel, then hash["ark_name"]
is a hash with the current ARK under the key "current".