Refactor is_retriable_exception logic
Just aiming to clarify the logic that decides whether a transaction
should be retried given an exception, but with one small change.

If a Java-level exception makes it to the top of a `DB.open` call,
Sequel wraps it in a `Sequel::DatabaseError` and rethrows.  This can
create pretty unhelpful error messages, because
`is_retriable_exception` was only expecting to see JDBC exceptions.
When that happens, you get something like:

     Error: #<NoMethodError: undefined method `getSQLState'; for java.lang.NullPointerException:Java::JavaLang::NullPointerException>

So, in addition to making the control flow a little more explicit,
this change also skips checking `getSQLState` on other exception
types.  They'll get re-raised without the confusing `NoMethodError`.