Farewell, embedded Solr (v4)!
ArchivesSpace has gotten a lot of mileage out of the embedded Solr.
This version of Solr is now very old and unsupported:

https://solr.apache.org/downloads.html#about-versions-and-support

To move forward we have to pull the plug on the embedded deployment:

https://solr.apache.org/news.html#apache-solr-500-and-reference-guide-for-50-available

**Solr will no longer ship a war file and instead be a downloadable application**

This PR removes the embedded Solr from ArchivesSpace. From this point
on it will be necessary to use an external Solr deployment, as is the
case for the database server (MySQL). This is true for both production
**AND** development. Documentation will be provided through Tech Docs.

The PR does not attempt to optimize or otherwise improve upon the
configuration that has been in place historically. Only the necessary
changes required to support the most recent Solr release have been made.