Change case/when/else to if/elsif/else
If don't know why, probably some subtlety of Ruby semantics, but the case/when/else in this method always executed the else-condition, regardless of the type passed to it. Hence the calling code in ASModel_crud.rb always got a false response, and so created a new ARK every time the record was updated. Replacing with the logically-identically if/elsif/else equivalent, which does allow the first two conditions to evaluate to true when the type passed to the method is either a Resource or ArchivalObject class.