Upgrade Jetty to v9
Includes jettyrunner and jettywebapp which is required when moving
up to Jetty 10. The updated launcher will work with v9 or v10.

To try with an existing release replace launcher.rb with this one:

https://gist.github.com/mark-cooper/53043a60603f516e89b654350160fa6d

That version restores support for the embedded Solr. Remove the jars
from launcher/lib & download the v9 or v10 jars to launcher/lib.
Tested with 3.0.2.