Remove unnecessary warning log triggered on concurrent logins
A few people on the listserv commented that they were seeing this
message a lot, due to integrations that were logging in as system
users multiple times concurrently.

The log message isn't really helpful anyway: it doesn't indicate a
problem and there's nothing to be done about it.  The comment in the
code explains what's happening, so I've dropped the logging.