Implements the fetch_resolved refactoring mentioned in the comment, by consolidating the method in application controller.
Also adds an excludes option to the method signature which allows for
an optimized way of loading accession & resource records by being more
selective about the data that is retrieved (vs. everything in find_opts).

Applies an exclusion for linked events (and their linked records) in
edit & update mode as that data is not required and can have a large
impact on performance when many linked event records are associated.