Performance improvement for loading of linked event records in SUI
Implements the fetch_resolved refactoring mentioned in the comment,
by consolidating the method in application controller.

Also adds an excludes option to the method signature which allows for
an optimized way of loading accession & resource records by being more
selective about the data that is retrieved (vs. everything in find_opts).

Applies an exclusion for linked events (and their linked records) in
edit & update mode as that data is not required and can have a large
impact on performance when many linked event records are associated.

Adds an AppConfig setting for max_linked_events_to_resolve. When
the max is exceeded (> no. of linked events) then the events and
linked records are not resolved in view mode and a simplified
display is presented (excludes titles). This keeps memory usage
under control as the no. of linked events grows.