Fixes ANW-1550 - replaces broken XMLCleaner with Nokogiri PushParser
The existing code, which is run per PDF finding aid generation on the
PUI, will run up to 10,000 times before failing in the face of invalid
XML with an error it doesn't understand, which is most XML errors.
This can take tens of minutes or more on large XML output.

The replacement code fixes all the same issues, and will fail early
and gracefully on the XML issues the previous code groveled over.