Fix wrong permission for creating/editing repositories
The backend API requires the "create_repository" global permission to create new repositories or edit the details (name, publish y/n, OAI settings, etc) of an existing one. Whereas "manage_repository" is a repository-level permission for controlling access to the current repository. This change hides buttons in the staff interface which lead to the form for editing a repository, but unless the user is a system admin they'll get a error if they try to save changes.