override sort preferences for linker search
The linker typeahead for agents and resources will now force a sort by score on the search results instead of pulling from the repo sort preferences. An optional parameter called "sorting" is added to Search::all for this purpose.

(initial title field query restriction was removed in favor of this)