ANW-1501: fix locations list holdings count cells
Should now properly select the location linker Browse button to add
the click handler, and populate the holdings counts cells both when
the modal is displayed and when the pagination controls are used.
MutationObservers are used both for ensuring we wait until the modal is
present and for detecting when new rows appear (on initial display and
on page change).