account for LargeTree client-side interactions
Resources and Archival Objects use this LargeTree Javascript
component for navigating and viewing. Since this happens entirely on
the client side, the record object that was provided to the view when
the user initially clicked Edit is no longer valid if the user clicks a
different record in the tree.

As a result, we needed some JavaScript that grabs the ID of the
"current" record (from the Tree) and tacks it on to the parameters
that are submitted by the linker's Create button. Now the controller
can copy fields from the correct record into the new Digital Object.

Also went back and fixed the name of the preference - not sure why
I had  "_description" on it.