Cleanup bugs preventing frontend feature tests from passing with a Chrome driver.
Switch to Puma webserver for frontend feature tests for now. (Not totally sure
it matters, but at least we stop seeing huge stacktraces when the browser
hangs up on the Jetty 9 webserver, and in any case Capybara seems to prefer
Puma as it is the default server.

Add some tuning to Capybara helpers to avoid errors arising from
slow JQuery event handlers not being attached in time for clicks

add a tiny bit of proper namespacing to ead importer

reconcile upgrades and master after rebase; upgrade Gemfiles