Ensure prefilled collection in "browse top containers" works with a URL prefix
When you click "Browse" from a Top Container linker, there's some JS
in linker.js that prepopulates the "Resource" field in the browse (to
show you top containers linked to the current Resource record).

This code relies on a regex on window.location.pathname to work out
whether it's running in the context of a resource record, but that
regex didn't account for the possibility of a URL prefix.  Modify that
regex to work when running under a prefix too.