ANW-1477 Password reset fixes/updates (#3049)
* allow backend to find templates on Jetty

The user_mailer uses standard ERB templates when composing emails. Being
that the backend is a Sinatra app, it doesn't have Action View available
for locating them. The method originally used to find them worked in the
dev environment, but not in a release running on Jetty. This minimal
change should append the correct view path in both cases. It will also
only append the path once (in the initializer) instead of doing it every
time send_reset_token is called.

* improve consistency of "reset password" language used in code and UI