login to backend regardless of factory usage
Establishing a backend session was being deferred until a factory
instance was created, which assumes all test suites will do so before
doing something that requires a backend session, such as running the
indexer. However, the accessibility test suite, which depends upon a
preloaded database, skips setting up test fixtures. In this case, that
means no factories are used, and therefore there is no backend session
available for the indexer to use, so setup fails with an
AccessDeniedException.

This change will log in during AspaceFactories initialization, so the backend
session will be created regardless of whether any factories are actually
used.