ANW-1693 Include family agents in PUI agent badge count (#2997)
The PUI agents listing includes agent records with the pui_agent_family
type, but the badge for them in the repository page ("Names" in English)
was not including that type in its count. This adds the type to the
list of counts that are fetched, and a special case for adding it in to
the pui_agent count for display.