add support for primary agent links to marc export
For exports: if there is a primary agent with role = creator, then this
agent gets the 1xx tag, with the rest getting a 7xx.

If there is no primary agent or a primary agent with a role != creator,
then behavior is unchanged, and the first creator agent gets the 1xx
tag.

For imports: The agent in the 1xx tag gets flagged as primary in the
import.