remove unused variable and fixup test
The js_path variable in InifiniteRecords was no longer needed
after the Collection Organization WIP load-all-records stuff
was backed out. As it stood it was resulting in a crash upon
visiting the CO page. The resources spec was also failing
a test that navigates to the CO page on a call to
finished_all_ajax_requests due to jQuery apparently not
being available on the CO page. The underlying cause(s)
may or may not need to be investigated later.