account for potential app proxy prefix in IR code
Passes in the public app proxy url to InfiniteTree/InfiniteRecords so
they can properly fetch nodes/waypoints when a proxy is in use.

This also ends up involving some changes so that the public URL can be
defined at the top level in build.xml and passed to rspec so that
AppConfig can be set appropriately and various tests involving the
collection organization pages will run successfully. Unfortunately, the
accessibilty CI tests run the rspec target directly instead of going
through their own setup target like other tests, so the url had to be
hardcoded in a parameter definition in the yaml file.