restore/refactor assets proxy prefix handling
Asset file handling for proxy prefixes was commented out of the
production.rb files at some point during the softserv work, and
though an updated version was restored to the staff side, the fix
appears to have been forgotten/missed for the PUI.

Additionaly, there was an older class called AssetPathRewriter that was
implemented to provide this functionality many years ago, and was never
removed. The impact appeared to be limited to just making copies of
asset files, but removing it seems to have no ill effect, so it's a nice
opportunity to clear a small amount of tech debt.

This has probably been noted before, but as this code is modifying
internal ActionDispatch functionality, it is liable to break again in
the future with Rails updates.