reproduce search "Found In" column in CSV exports
The "Found In" column in the search results table in the Staff UI
is generated by the frontend, and unlike the other columns is not
indexed so it doesn't show up in CSV exports. This is because the CSV
export helper actually has to re-run the search in order to get the full
results set (not just the currently displayed page), and is just given
the field names displayed in the results table to include in the search.
Since "context" is not an actual field in any record, it comes back
blank.

I've added a class to the ExportHelper that mimics the way the
"Found In" column results are constructed. Since the various fields that
indicate the ancestors for different record types only provide URI refs,
they need to be grabbed from the backend in order to get the actual
title string. This introduces a lot of overhead, so the titles are
cached as they are retreived to avoid having to make this a background
job. Since most ancestors are likely to be parents to many of the
records in the search results, this should be effective.