restore and correct AssetPathRewriter
I removed this class before because I thought it was redundant to some
other newer stuff in production.rb that is handling asset prefixes,
but as it turns out, it was not. Also updated it to look for CSS files
anywhere under the assets directory so future additions will hopefully
be handled.