fix empty path segment in top containers search
Jetty 10 apparently decided to start complaining about "empty segments"
in URIs and returning 400 errors when they are encountered. It often
happens that URIs end up with double slashes in them when being
constructed via string interpolation, and this is one spot where that
was happening. We'll need to be more careful about that going forward.