smash the building location hierarchy
The location model is stored in a non-hierarchical fashion, with any
variation in field values resulting in a new database row. The only
required field is Building, so users can input locations that have only
a Building and a Room, among other permutations.

However, the location model's building_data method was organizing the
flat rows from the database into a hierarchy of building->floor->
room->area. If a floor did not have any rooms defined for it, it would
move on to the next location row, thus not allowing for rooms without
floors (same for areas without rooms).

Since all buildings have at least one floor, it would be ideal for the
location creation form to require the floor field as well. However,
as a free text field, this could result in some users entering "1" while
others enter "first" or the like. Also, one can imagine a floor of a
building not having any rooms, but having areas (perhaps a warehouse),
so it makes sense not to require the room field, which also breaks the
hierarchy.

This change to the building_data method continues to process rooms
even if there is no floor, and likewise areas even if there is no room,
using simple strings to represent these cases.