fix event +1 redirecting to edit new event
The previous change to fix the event +1 button causing a crash when
coming from the main menu Create->Event button was erroneously
redirecting to editing the newly created event; it should redirect to
a blank new event page.

This should be cleaned up now, and I've left copious comments
because this was frustratingly confusing to me and I'd like to spare
future developers from that.