Gem group exclusion broke when bundler was upgraded
The issue seems to be the bundler makes assumptions about
the location of the application directory that resulted in
the config files being in the wrong place: when the config
was being written, bundler assumed the root of the repository
was the application home, but then when the various Gemfiles
are installed, bundler assumes the location of each Gemfile is
the application home. So this commit hardcodes the app home to
always be `build/.bundle`.