Update aspace cookies
Set secure when in production (requires https)
Set SameSite lax (Rails default from v6 (?), and will be a browser
default soon-ish ~ at least with Firefox). Strict will be a problem
for sites who use different domains for staff & public ui.
Set httponly when not js (it already was, but this makes it clear)