Use $gemfile param to set .bundle in the same directory.
Revision of previous commit. It turns out to be a bad idea to
try to use a single .bundle/config for all the applications. The
.bundle/config files need to persist in the wars so that statements
like `require 'bundler/setup' don't error when they can't find gems
that were omitted during the build.
Since in practice every call to the "bundler" task includes a
`gemfile` param with a value like {path}/Gemfile, we can just use sed
to capture the path and thus make sure every `bundle config` command
within the task will be persisted to the same directory as the
Gemfile.