Revert "Use the hash syntax when translating enumeration values in the listing."
This reverts commit 204c48109a994c6a99ef743b2c6d0abf5197b712.
The commit was submitted in PR 301:
https://github.com/archivesspace/archivesspace/pull/301
There was no explanation other than supporting I18n keys
with '.' in the string. I could find no examples with a '.'
in the key in any of the enumerations in the ArchivesSpace
repository, and this alternate way of calling `translate`
appears not to work in current versions of I18n. So while
previous versions of ArchivesSpace do show translated values
correctly in the controlled value list table, v4.0.0-RC1 does
not. See previous commit for test.