Remove Required Field Templates for Notes
Once upon a time there was a line like this in `agents/form_required':

<%= render_aspace_partial :partial => "notes/form_required", :locals => {:form => form, :all_note_types => note_types_for(form['jsonmodel_type']), :section_id => "#{@agent.agent_type}_notes"} %>

But it was removed a long time ago when a raft of changes to the
Agents module was merged:
https://github.com/archivesspace/archivesspace/pull/1804

So it should be safe to dispense with these now. There do not appear
to be any note required options on the required field form for agents.