Remove dynamic _relator_sort fields from solr schema.
Many moons ago there was a PR which introduced some PUI functionality
which apparently depended or was though to depend on being able to
sort a record's related agents by their role.
https://github.com/archivesspace/archivesspace/pull/429
None of this appears to be in use or ever to have been in use in the
public app, but the indexer is still creating dynamic fields with
keys like:

_agents_corporate_entities_102_relator_sort
_agents_corporate_entities_103_relator_sort
_agents_corporate_entities_104_relator_sort

In fact, for an instance with 1000 agent records, the Solr schema will
contain 1000 of these unused, unecessary fields in the dynamic schema.
(Try it by looking at the "Schema" pane in the Solr admin panel.)
I don't know how much this has been affecting Solr performance, but
removing all these spam fields can surely only help.