delete uri from revision_statement schema
Revision Statements appear to be basic subrecords like
extents, dates, etc. It appears that 'uri' was added to the
json schema in error. This causes an obscure bug where if a
job creates a revision statement, it will try to register it
as a top-level record and will error when trying to display
the list of job-created records.