fixes missing title error message
This is actually caused by the required_subrecord_form partial (line 24)
calling wrap_with_tooltip, which in turn does not account for the
possibility that there may not be a tooltip for the subrecord type
(as was the case with titles, since I hadn't added one yet). This
results in the "subrecord-form-heading-label" class not being applied,
which breaks the client-side (!) code that inserts the appropriate field
name and the little chevron that links down to the associated subrecord.

For now, I've updated the old title tooltip to now apply to the
multiple titles subrecord, which alleviates the issue. Not sure if this
is something that warrants additional work right now.