Avoid Memory Leaks in Converters
When a JSONModel instance's `id` method is invoked,
calls `parse_reference`, which uses a cache mapping uri
patterns to models, e.g. /respositories/archival_objects =>
['arhival_object, <Class xxxxx>]
This is problematic if the object happens to have a non-standard
uri pattern, as is the case during imports. In these cases
an entry was being added to the cache everytime `id` was invoked
on any object during an import conversion.